backend/services/query_processor.py
from typing import Dict, List
import spacy

class QueryProcessor:
    def __init__(self):
        self.nlp = spacy.load("en_core_web_sm")
        
    def process_query(self, query: str) -> Dict:
        """
        Process natural language shopping query
        """
        doc = self.nlp(query)
        
        # Extract key information
        price_range = self._extract_price_range(doc)
        product_type = self._extract_product_type(doc)
        attributes = self._extract_attributes(doc)
        
        return {
            "price_range": price_range,
            "product_type": product_type,
            "attributes": attributes
        }
    
    def _extract_price_range(self, doc) -> Dict:
        # TODO: Implement price range extraction
        pass
    
    def _extract_product_type(self, doc) -> str:
        # TODO: Implement product type extraction
        pass
    
    def _extract_attributes(self, doc) -> List[str]:
        # TODO: Implement attribute extraction
        pass 